feat: standardized-provider-migrate-gemini-007 - Close Gemini migration with focused and quality-gate verification
