+++ docs/internal/baselines/ownership-inventory.json
@@ -3674,6 +3674,14 @@
       "deletionCondition": "delete Workers provider packages after Providers root cutover proof (IMP-providers-*)"
     },
     {
+      "packagePath": "pkg/services/workers/provider/conductor",
+      "disposition": "move",
+      "destination": "providers",
+      "destinationKind": "owner",
+      "successor": "pkg/services/providers",
+      "deletionCondition": "delete Workers provider packages after Providers root cutover proof (IMP-providers-*)"
+    },
+    {
       "packagePath": "pkg/services/workers/provider/cursor",
       "disposition": "move",
       "destination": "providers",
