+++ tests/functional/workers/gemini/process_harness_test.go
